+++ plugin/clang/utils.py
@@ -13,86 +13,6 @@ from os import path
 log = logging.getLogger("ECC")
 
 
-class MacroParser(object):
-    """Parse info from macros.
-
-    Clang doesn't provide much information for MACRO_DEFINITION cursors,
-    so we have to parse this info ourselves.
-    """
-
-    def __init__(self, name, location):
-        """Parse the macro with the given name from its location.
-
-        Args:
-            name (str): Macro's name.
-            location (cindex.SourceLocation): Macro definition's location.
-                Can be None for unittests that parse with text content
-                directly instead of loading from file.
-
-        Future improvement: if there are line continuation characters
-            in a macro with parenthesis, continue parsing into the next line
-            to find it and create a proper args string.
-        """
-        self._args_string = ''
-        self._name = name
-        self._body = ''
-        if location and location.file and location.file.name:
-            with open(location.file.name, 'r', encoding='utf-8',
-                      errors='ignore') as f:
-                macro_file_lines = f.readlines()
-                self._parse_macro_file_lines(macro_file_lines, location.line)
-
-    def _parse_macro_file_lines(self, macro_file_lines, macro_line_number):
-        """Parse a macro from lines of text containing the macro.
-
-        Args:
-            macro_file_lines (list[str]): lines of text containing the macro
-            macro_line_number (int): line number (1-based) of the macro
-               in macro_file_lines.
-        """
-        macro_line = macro_file_lines[macro_line_number - 1].strip()
-        # strip leading '#<whitespace>define<whitespace><macro name>'
-        macro_line = macro_line.lstrip('#').lstrip().lstrip('define')
-        macro_line = macro_line.lstrip().lstrip(self._name)
-        # macro that looks like a function, possibly with args
-        if macro_line.startswith('('):
-            end_args_index = macro_line.find(')')
-            # There should always be a closing parenthesis, but check
-            # just in case a) the code is malformed or b) the macro
-            # definition is continued on the next line so we can't
-            # find it on this line.
-            if end_args_index != -1:
-                # If extra spaces, e.g. "#define MACRO( x ,  y  , z )",
-                # then flatten down to just "(x, y, z)"
-                args_str = macro_line[1:end_args_index]
-                args_str = ''.join(args_str.split())
-                args_str = args_str.replace(',', ', ')
-                self._args_string = '(' + args_str + ')'
-                macro_line = macro_line[end_args_index + 1:]
-
-        self._body = macro_line.strip()
-        while self._body.endswith("\\"):
-            macro_line_number += 1
-            line = macro_file_lines[macro_line_number - 1].rstrip()
-            self._body += "\n" + line
-
-    @property
-    def args_string(self):
-        """Get arguments string.
-
-        Examples:
-            '#define MACRO()' would return '()'
-            '#define MACRO(x, y)' would return '(x, y)'
-            '#define MACRO' would return ''
-        """
-        return self._args_string
-
-    @property
-    def body_string(self):
-        """Get macro body string."""
-        return self._body
-
-
 class ClangUtils:
     """Utils to help handling libclang, e.g. searching for it.
 
@@ -233,50 +153,6 @@ class ClangUtils:
         return None
 
     @staticmethod
-    def location_from_type(clangType):
-        """Return location from type.
-
-        Return proper location from type.
-        Remove all inderactions like pointers etc.
-
-        Args:
-            clangType (cindex.Type): clang type.
-
-        """
-        cursor = clangType.get_declaration()
-        if cursor and cursor.location and cursor.location.file:
-            return cursor.location
-
-        cursor = clangType.get_pointee().get_declaration()
-        if cursor and cursor.location and cursor.location.file:
-            return cursor.location
-
-        return None
-
-    @staticmethod
-    def link_from_location(location, text):
-        """Provide link to given cursor.
-
-        Transforms SourceLocation object into html string.
-
-        Args:
-            location (Cursor.location): Current location.
-            text (str): Text to be added as info.
-        """
-        result = ""
-        if location and location.file and location.file.name:
-            result += "<a href=\""
-            result += location.file.name
-            result += ":"
-            result += str(location.line)
-            result += ":"
-            result += str(location.column)
-            result += "\">" + text + "</a>"
-        else:
-            result += text
-        return result
-
-    @staticmethod
     def get_text_by_extent(extent):
         """Load lines of code in range, pointed by extent.
 
@@ -305,121 +181,6 @@ class ClangUtils:
 
         # sublime does not recognize &quot;
         result = result.replace("&quot;", "&nbsp;")
-        return result
-
-    @staticmethod
-    def build_info_details(cursor, cindex, settings):
-        """Provide information about given cursor.
-
-        Builds detailed information about cursor.
-
-        Args:
-            cursor (Cursor): Current cursor.
-            cindex (module): clang cindex.py module for the correct version
-
-        """
-        type_decl = [
-            cindex.CursorKind.STRUCT_DECL,
-            cindex.CursorKind.UNION_DECL,
-            cindex.CursorKind.CLASS_DECL,
-            cindex.CursorKind.ENUM_DECL,
-            cindex.CursorKind.TYPEDEF_DECL,
-            cindex.CursorKind.CLASS_TEMPLATE,
-            cindex.CursorKind.TYPE_ALIAS_DECL,
-            cindex.CursorKind.TYPE_REF
-        ]
-
-        result = ""
-
-        result += '<b>Declaration:</b><br>'
-
-        # Show the return type of the function/method if applicable,
-        # macros just show that they are a macro.
-        macro_parser = None
-        is_macro = cursor.kind == cindex.CursorKind.MACRO_DEFINITION
-        is_type = cursor.kind in type_decl
-        if is_macro:
-            macro_parser = MacroParser(cursor.spelling, cursor.location)
-            result += '#define '
-        else:
-            if cursor.result_type.spelling:
-                result_type = cursor.result_type
-            elif cursor.type.spelling:
-                result_type = cursor.type
-            else:
-                result_type = None
-                log.warning("No spelling for type provided in info.")
-                return ""
-
-            if cursor.is_static_method():
-                result += "static "
-
-            if cursor.spelling != cursor.type.spelling:
-                # Don't show duplicates if the user focuses type, not variable
-                result += ClangUtils.link_from_location(
-                    ClangUtils.location_from_type(result_type),
-                    html.escape(result_type.spelling))
-                result += ' '
-
-        # Link to declaration of item under cursor
-        if cursor.location:
-            result += ClangUtils.link_from_location(
-                cursor.location,
-                html.escape(cursor.spelling))
-        else:
-            result += html.escape(cursor.spelling)
-
-        # Macro/function/method arguments
-        args_string = None
-        if is_macro:
-            # cursor.get_arguments() doesn't give us anything for macros,
-            # so we have to parse those ourselves
-            args_string = macro_parser.args_string
-        else:
-            args = []
-            for arg in cursor.get_arguments():
-                if arg.spelling:
-                    args.append(arg.type.spelling + ' ' + arg.spelling)
-                else:
-                    args.append(arg.type.spelling + ' ')
-            if cursor.kind in [cindex.CursorKind.FUNCTION_DECL,
-                               cindex.CursorKind.CXX_METHOD]:
-                args_string = '('
-                if len(args):
-                    args_string += html.escape(', '.join(args))
-                args_string += ')'
-        if args_string:
-            result += args_string
-
-        # Show value for enum
-        if cursor.kind == cindex.CursorKind.ENUM_CONSTANT_DECL:
-            result += " = " + str(cursor.enum_value)
-            result += "(" + hex(cursor.enum_value) + ")"
-
-        # Method modifiers
-        if cursor.is_const_method():
-            result += " const"
-
-        # Show macro body
-        if is_macro:
-            result += " " + ClangUtils.htmlize_text(macro_parser.body_string)
-
-        # Show type declaration
-        if settings.show_type_body and is_type and cursor.extent:
-            body = ClangUtils.get_text_by_extent(cursor.extent)
-            result += "<br><br>" + ClangUtils.htmlize_text(body)
-
-        # Doxygen comments
-        if cursor.brief_comment:
-            result += "<br><br><b>Brief documentation:</b><br>"
-            result += "<div>" + cursor.brief_comment + "</div>"
-
-        if cursor.raw_comment:
-            html_raw_comment = ClangUtils.cleanup_comment(cursor.raw_comment)
-            if len(html_raw_comment) > 0:
-                result += "<br><b>Full Doxygen comment:</b>"
-                result += "<br><div>" + html_raw_comment + "</div>"
-
         return result
 
     @staticmethod
