--- plugin/error_vis/popup_error.md
@@ -1,4 +1,0 @@
-!!! {type}
-    ```bash
-    {content}
-    ```
