+++ plugin/error_vis/popup_error_vis.py
@@ -4,15 +4,13 @@ Attributes:
     log (logging): this module logger
 """
 import logging
-import mdpopups
 import sublime
 from os import path
-from string import Template
-
-import sublime
 
 from ..completion.compiler_variant import LibClangCompilerVariant
 from ..settings.settings_storage import SettingsStorage
+from ..popups.popups import Popup
+from ..popups.popups import PopupStyle
 
 log = logging.getLogger("ECC")
 
@@ -20,11 +18,6 @@ PATH_TO_HTML_FOLDER = path.join(
     path.dirname(path.dirname(__file__)), 'html')
 
 PATH_TO_ICON = "Packages/EasyClangComplete/pics/icons/{icon}"
-
-POPUP_ERROR_HTML_FILE = path.join(PATH_TO_HTML_FOLDER, "error_popup.html")
-POPUP_WARNING_HTML_FILE = path.join(PATH_TO_HTML_FOLDER, "warning_popup.html")
-
-POPUP_MD_FILE = "Packages/EasyClangComplete/plugin/error_vis/popup_error.md"
 
 
 class PopupErrorVis:
@@ -34,14 +27,8 @@ class PopupErrorVis:
         err_regions (dict): dictionary of error regions for view ids
     """
     _TAG = "easy_clang_complete_errors"
-    _MAX_POPUP_WIDTH = 1800
     _ERROR_FLAGS = sublime.DRAW_EMPTY | sublime.DRAW_NO_FILL
     _ERROR_SCOPE = "invalid.illegal"
-
-    ERROR_HTML_TEMPLATE = Template(
-        open(POPUP_ERROR_HTML_FILE, encoding='utf8').read())
-    WARNING_HTML_TEMPLATE = Template(
-        open(POPUP_WARNING_HTML_FILE, encoding='utf8').read())
 
     def __init__(self, gutter_style=None):
         """Initialize error visualization.
@@ -153,28 +140,14 @@ class PopupErrorVis:
         if row in current_err_region_dict:
             errors_dict = current_err_region_dict[row]
             max_severity, error_list = PopupErrorVis._as_list(errors_dict)
-            popup_type = 'panel-warning "ECC: Warning"'
 
-            css = '''
-.ECC .admonition-title
-{
-    border-top-left-radius: 0rem;
-    border-top-right-radius: 0rem;
-}
-.ECC .admonition
-{
-    border-radius: 0rem;
-}
-'''
+            text_to_show = '\n'.join(error_list)
+            popup = None
             if max_severity > 2:
-                popup_type = 'panel-error "ECC: Error"'
-            text = sublime.load_resource(POPUP_MD_FILE).format(
-                type=popup_type, content='\n'.join(error_list))
-            mdpopups.show_popup(view, text, max_width=self._MAX_POPUP_WIDTH,
-                                wrapper_class='ECC',
-                                css=css)
-            # errors_html = PopupErrorVis._as_html(errors_dict)
-            # view.show_popup(errors_html, max_width=self._MAX_POPUP_WIDTH)
+                popup = Popup.error(text_to_show)
+            else:
+                popup = Popup.warning(text_to_show)
+            popup.show(view)
         else:
             log.debug("no error regions for row: %s", row)
