+++ plugin/popups/popup.css
@@ -1,0 +1,11 @@
+.ECC .admonition-title
+{
+    border-radius: 0rem;
+    padding-left: 10px;
+}
+.ECC .admonition
+{
+    border-radius: 0rem;
+    padding: 0rem;
+    top: -3rem;
+}
