+++ plugin/popups/popup.md
@@ -1,0 +1,2 @@
+!!! {type}
+    {contents}
